feat: add better structural organization for rewards contracts
