+++ packages/react-app-revamp/lib/rewards/contracts/batch-operations.ts
@@ -1,0 +1,56 @@
+import { config } from "@config/wagmi";
+import { readContracts } from "@wagmi/core";
+import { Abi } from "viem";
+import { ModuleType } from "../types";
+
+/**
+ * Creates a contract read batch for native token data
+ */
+export const createNativeTokenReadBatch = (
+  moduleType: ModuleType,
+  contractAddress: `0x${string}`,
+  chainId: number,
+  abi: Abi,
+  validRankings: number[],
+  voterAddress: `0x${string}`,
+  functionName: string,
+) => {
+  return readContracts(config, {
+    contracts: validRankings.map(ranking => ({
+      address: contractAddress,
+      chainId,
+      abi,
+      functionName,
+      args: moduleType === ModuleType.VOTER_REWARDS ? [voterAddress, BigInt(ranking)] : [BigInt(ranking)],
+    })),
+  });
+};
+
+/**
+ * Creates a contract read for ERC20 token data
+ */
+export const createERC20TokenRead = (
+  moduleType: ModuleType,
+  contractAddress: `0x${string}`,
+  chainId: number,
+  abi: Abi,
+  tokenAddress: `0x${string}`,
+  ranking: number,
+  voterAddress: `0x${string}`,
+  functionName: string,
+) => {
+  return readContracts(config, {
+    contracts: [
+      {
+        address: contractAddress,
+        chainId,
+        abi,
+        functionName,
+        args:
+          moduleType === ModuleType.VOTER_REWARDS
+            ? [tokenAddress, voterAddress, BigInt(ranking)]
+            : [tokenAddress, BigInt(ranking)],
+      },
+    ],
+  });
+};
