+++ packages/react-app-revamp/lib/rewards/contracts/validation.ts
@@ -1,0 +1,76 @@
+import { config } from "@config/wagmi";
+import { readContracts } from "@wagmi/core";
+import { compareVersions } from "compare-versions";
+import { Abi } from "viem";
+import { VOTER_REWARDS_VERSION } from "../types";
+
+/**
+ * Validates rankings
+ */
+export const validateRankings = async (
+  rankings: number[],
+  contractAddress: `0x${string}`,
+  chainId: number,
+  abi: Abi,
+  version: string,
+  creatorAddress?: `0x${string}`,
+): Promise<{ validRankings: number[]; tiedRankings: number[] }> => {
+  if (!rankings.length) return { validRankings: [], tiedRankings: [] };
+
+  // check if this is an older version (before VOTER_REWARDS_VERSION)
+  if (compareVersions(version, VOTER_REWARDS_VERSION) < 0) {
+    // for older versions, use getAddressToPayOut
+    const addressResults = await readContracts(config, {
+      contracts: rankings.map(ranking => ({
+        address: contractAddress,
+        chainId,
+        abi,
+        functionName: "getAddressToPayOut",
+        args: [BigInt(ranking)],
+      })),
+    });
+
+    const validRankings: number[] = [];
+    const tiedRankings: number[] = [];
+
+    for (let i = 0; i < rankings.length; i++) {
+      const payoutAddress = addressResults[i]?.result as `0x${string}` | undefined;
+      if (payoutAddress !== undefined) {
+        //TODO: if creator participated, it isn't a tie, just think about this
+        if (payoutAddress.toLowerCase() === creatorAddress?.toLowerCase()) {
+          tiedRankings.push(rankings[i]);
+        } else {
+          validRankings.push(rankings[i]);
+        }
+      }
+    }
+
+    return { validRankings, tiedRankings };
+  }
+
+  const proposalIdResults = await readContracts(config, {
+    contracts: rankings.map(ranking => ({
+      address: contractAddress,
+      chainId,
+      abi,
+      functionName: "getProposalIdOfRanking",
+      args: [BigInt(ranking)],
+    })),
+  });
+
+  const validRankings: number[] = [];
+  const tiedRankings: number[] = [];
+
+  for (let i = 0; i < rankings.length; i++) {
+    const proposalId = proposalIdResults[i]?.result as bigint | undefined;
+    if (proposalId !== undefined) {
+      if (proposalId === 0n) {
+        tiedRankings.push(rankings[i]);
+      } else {
+        validRankings.push(rankings[i]);
+      }
+    }
+  }
+
+  return { validRankings, tiedRankings };
+};
